update 100% corrected answer

+++ Interview/AlgorithmTesterProject/AlgorithmTesterProject/WinForm/Form1.cs
@@ -18,50 +18,10 @@ namespace WinForm
         }
 
         private void Form1_Load(object sender, EventArgs e)
-        {   
+        {
             int[] A = new int[] { -1, 3, -4, 5, 1, -6, 2, 1 };
             solution(A);
         }
-
-        /// <summary>
-        /// equilibrium index
-        /// </summary>
-        /// <param name="A"></param>
-        /// <returns></returns>
-        //public int[] solution(int[] A)
-        //{
-        //    List<int> List_ReturnInt = new List<int>();
-        //    if (A.Length <= 0) {
-        //        List_ReturnInt.Add(-1);
-        //        return List_ReturnInt.ToArray(); 
-        //    }
-
-        //    int LeftSumValue = 0;
-        //    int RightSumValue = 0;
-            
-        //    /// Get right sum total.
-        //    for (int i = 0; i < A.Length; i++)
-        //    {
-        //        RightSumValue += A[i];
-        //    }
-
-        //    /// Get left value and compare
-        //    for (int i = 0; i < A.Length; i++) 
-        //    {
-        //        int CurrentValue = RightSumValue - A[i];
-        //        if (LeftSumValue == CurrentValue)
-        //        {
-        //            /// found equilibrium
-        //            List_ReturnInt.Add(i);
-        //        }
-        //        else 
-        //        {
-        //            LeftSumValue += A[i];
-        //            RightSumValue = CurrentValue;
-        //        }
-        //    }
-        //    return List_ReturnInt.ToArray();
-        //}
 
         public int solution(int[] A)
         {   
@@ -70,8 +30,8 @@ namespace WinForm
                 return -1;
             }
 
-            int LeftSumValue = 0;
-            int RightSumValue = 0;
+            long LeftSumValue = 0;
+            long RightSumValue = 0;
 
             /// Get right sum total.
             for (int i = 0; i < A.Length; i++)
@@ -81,8 +41,8 @@ namespace WinForm
 
             /// Get left value and compare
             for (int i = 0; i < A.Length; i++)
-            {
-                int CurrentValue = RightSumValue - A[i];
+            {                
+                long CurrentValue = RightSumValue - A[i];
                 if (LeftSumValue == CurrentValue)
                 {
                     /// found equilibrium
